pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-git-versioning<2"]
build-backend = "setuptools.build_meta"

[project]
name = "luna-soc"
description = "Amaranth HDL libary for building SoCs with LUNA USB peripherals."
license = { text = "BSD" }
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Great Scott Gadgets", email = "dev@greatscottgadgets.com"},
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Environment :: Console",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Topic :: System :: Hardware :: Universal Serial Bus (USB)",
]

dynamic = ["version"]

dependencies = [
    "luna-usb @ git+https://github.com/greatscottgadgets/luna@main",
]

[project.urls]
Documentation = "https://luna-soc.readthedocs.io"
Repository    = "https://github.com/greatscottgadgets/luna-soc"
Issues        = "https://github.com/greatscottgadgets/luna-soc/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
luna_soc = "luna_soc"

[tool.setuptools.package-data]
"luna_soc.gateware.cpu.verilog.vexriscv" = ["*.v"]

[tool.setuptools-git-versioning]
enabled = true
starting_version = "0.0.0"